Don't pollute Duplicate_Of list
Previously the updating list of duplicate issues would contain the total set of
issues which are marked as duplicates of the current issue. Thus is the metadata
was updated the issue would then have all those issues marked as duplicates.
This is incorrect as it pollutes the data the user has entered with the computed
set of duplicates. Fix this by only having the issue specific metadata shown in
the Duplicate_Of metadata update box.


git-svn-id: svn+ssh://svnrepo/projects/nitpick/trunk@5178 561c3618-0fe6-0310-b5d6-fc65c73fee30